// src/lib/list/enums/focus-alignments.ts
/**
 * Focus Alignments.
 * @link https://github.com/DjonnyX/centrifugal/blob/main/src/lib/list/enums/focus-alignments.ts
 * @author Evgenii Alexandrovich Grebennikov
 * @email djonnyx@gmail.com
 */
export enum FocusAlignments {
    NONE = 'none',
    START = 'start',
    CENTER = 'center',
    END = 'end',
}