// src/lib/list/enums/selecting-modes-types.ts
/**
 * Mods for selecting list items.
 * @link https://github.com/DjonnyX/centrifugal/blob/main/src/lib/list/enums/selecting-modes-types.ts
 * @author Evgenii Alexandrovich Grebennikov
 * @email djonnyx@gmail.com
 */
export enum SelectingModesTypes {
    /**
     * List items are not selectable.
     */
    NONE = 0,
    /**
     * List items are selected one by one.
     */
    SELECT = 1,
    /**
     * Multiple selection of list items.
     */
    MULTI_SELECT = 2,
}